Add `door` control using 2 GPIO pins.
This is the main function of this project: Control a shutter over
MQTT and properly track the state of it through door sensors. Externally
the state is published and commands can be set through MQTT. Internally
this uses 4 GPIO pins to drive a motor fw/rev relay board (2) and door
sensors (2).