Functional load enable/disable.
The dimming value doesn't seem to be used. I also noticed my
controller was set to 5min relay 10V Load value, and 6.6A LED
load current setting, whatever that actually means.

Many of the bits in these registers are documented in a way that
they indicate that they influence how the load is controlled, but
not in an unambiguous way unfortunately.